+++ NEWS.md
@@ -1,8 +1,12 @@
 # lgr 0.1.1.9000
 
+* `get_loggers()` registeres new loggers in the lgr::loggers namespace, this 
+  is a more global and decoupled approach to loggers
 * removed `full_name` active binding for loggers. Loggers now only have 
   qualified names and `name` is now identical to what `full_name` was before.
   For consistency the format method of `ancestry` has also been revised.
+* New Loggers now inherit the log level of their parent unless one is set
+
 
 # lgr 0.1.1